move heatmap's `Worker` under `offline`
