+++ dsi/offline/heatmap/worker.py
